Refactor code to remove locking
