+++ README.md
@@ -728,6 +728,7 @@ This section contains detailed documentation for the following aspects of Terrag
 
 1. [Helper functions](#helper-functions)
 1. [CLI options](#cli-options)
+1. [Migrating from Terragrunt v0.11.x and Terraform 0.8.x and older](#migrating-from-terragrunt-v0.11.x-and-terraform-0.8.x-and-older)
 1. [Terragrunt config files](#terragrunt-config-files)
 1. [Developing Terragrunt](#developing-terragrunt)
 1. [License](#license)
@@ -912,6 +913,113 @@ To migrate, all you need to do is:
 1. Paste those contents into a `terragrunt = { ... }` block in a `terraform.tfvars` file.
 1. Delete the `.terragrunt` file.
 
+
+### Migrating from Terragrunt v0.11.x and Terraform 0.8.x and older
+ 
+Terraform 0.8.x and older did not support:
+
+1. Defining remote state configuration in a file
+1. Locking
+
+Terragrunt was originally created to support both of these features. As of Terraform 0.9.0, both of these features are
+supported natively, so we have had to make changes to Terragrunt: 
+
+1. Terragrunt still supports remote state configuration so you can take advantage of Terragrunt's interpolation
+   functions.
+1. Terragrunt no longer supports locking.
+
+If you were using Terragrunt v0.11.x and Terraform 0.8.x or older, here is how to migrate:
+
+1. You may leave your `remote_state` configuration largely unchanged. However, in your Terraform code, you must now
+   define a `backend`, although you can leave its configuration blank. For example, to use S3 as a remote state 
+   backend, you will need to add the following to your Terraform code:
+   
+    ```hcl
+    terraform {
+      backend "s3" {}
+    }
+    ```
+   
+    You can continue to use Terragrunt to configure this remote state backend in your `terraform.tfvars` files, taking 
+    advantage of Terragrunt's interpolation functions:
+    
+    ```hcl
+    terragrunt = {
+      remote_state {
+        backend = "s3"
+        config {
+          bucket     = "my-terraform-state"
+          key        = "${path_relative_to_include()}/terraform.tfstate"
+          region     = "us-east-1"
+          encrypt    = true
+        }
+      }
+    }
+    ```    
+   
+1. Remove any `lock { ... }` blocks from your Terragrunt configurations, as these are no longer supported. If you were
+   storing remote state in S3 and relying on DynamoDB as a locking mechanism, Terraform now supports that natively. 
+   To enable it, simply add the `lock_table` parameter to your S3 backend configuration. If you configure your S3 
+   backend using Terragrunt, then Terragrunt will automatically create the `lock_table` for your if it doesn't already
+   exist:
+
+    ```hcl
+    terragrunt = {
+      remote_state {
+        backend = "s3" 
+        config {
+          bucket     = "my-terraform-state"
+          key        = "${path_relative_to_include()}/terraform.tfstate"
+          region     = "us-east-1"
+          encrypt    = true
+       
+          # Tell Terraform to do locking using DynamoDB. Terragrunt will automatically create this table for you if
+          # it doesn't already exist. 
+          lock_table = "my-lock-table"
+        }
+      }
+    }
+    ```
+
+    If you would like Terraform to automatically retry locks like Terragrunt did (this is particularly useful when 
+    running Terraform as part of an automated script, such as a CI build), you use an `extra_arguments` block:
+
+    ```hcl
+    terragrunt = {
+      remote_state {
+        backend = "s3" 
+        config {
+          bucket     = "my-terraform-state"
+          key        = "${path_relative_to_include()}/terraform.tfstate"
+          region     = "us-east-1"
+          encrypt    = true
+    
+          # Tell Terraform to do locking using DynamoDB. Terragrunt will automatically create this table for you if
+          # it doesn't already exist. 
+          lock_table = "my-lock-table"
+        }
+      }
+    }
+    
+    # Force Terraform to keep trying to acquire a lock for up to 20 minutes if someone else already has the lock 
+    extra_arguments "retry_lock" {
+      commands = [
+        "init",
+        "apply",
+        "refresh",
+        "import",
+        "plan",
+        "taint",
+        "untaint"
+      ]
+    
+      arguments = [
+        "-lock-timeout=20m"
+      ]    
+    }   
+    ``` 
+
+ 
 
 ### Developing terragrunt
