+++ dynamodb/dynamo_lock_table.go
@@ -9,11 +9,32 @@ import (
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/util"
 	"time"
+	"github.com/gruntwork-io/terragrunt/aws_helper"
 )
 
 // DynamoDB only allows 10 table creates/deletes simultaneously. To ensure we don't hit this error, especially when
 // running many automated tests in parallel, we use a counting semaphore
 var tableCreateDeleteSemaphore = NewCountingSemaphore(10)
+
+// Terraform requires the DynamoDB table to have a primary key with this name
+const ATTR_LOCK_ID = "LockID"
+
+// Default is to retry for up to 5 minutes
+const MAX_RETRIES_WAITING_FOR_TABLE_TO_BE_ACTIVE = 30
+const SLEEP_BETWEEN_TABLE_STATUS_CHECKS = 10 * time.Second
+
+const DEFAULT_READ_CAPACITY_UNITS = 1
+const DEFAULT_WRITE_CAPACITY_UNITS = 1
+
+// Create an authenticated client for DynamoDB
+func CreateDynamoDbClient(awsRegion, awsProfile string) (*dynamodb.DynamoDB, error) {
+	session, err := aws_helper.CreateAwsSession(awsRegion, awsProfile)
+	if err != nil {
+		return nil, err
+	}
+
+	return dynamodb.New(session), nil
+}
 
 // Create the lock table in DynamoDB if it doesn't already exist
 func CreateLockTableIfNecessary(tableName string, client *dynamodb.DynamoDB, terragruntOptions *options.TerragruntOptions) error {
@@ -53,11 +74,11 @@ func CreateLockTable(tableName string, readCapacityUnits int, writeCapacityUnits
 	terragruntOptions.Logger.Printf("Creating table %s in DynamoDB", tableName)
 
 	attributeDefinitions := []*dynamodb.AttributeDefinition{
-		&dynamodb.AttributeDefinition{AttributeName: aws.String(ATTR_STATE_FILE_ID), AttributeType: aws.String(dynamodb.ScalarAttributeTypeS)},
+		&dynamodb.AttributeDefinition{AttributeName: aws.String(ATTR_LOCK_ID), AttributeType: aws.String(dynamodb.ScalarAttributeTypeS)},
 	}
 
 	keySchema := []*dynamodb.KeySchemaElement{
-		&dynamodb.KeySchemaElement{AttributeName: aws.String(ATTR_STATE_FILE_ID), KeyType: aws.String(dynamodb.KeyTypeHash)},
+		&dynamodb.KeySchemaElement{AttributeName: aws.String(ATTR_LOCK_ID), KeyType: aws.String(dynamodb.KeyTypeHash)},
 	}
 
 	_, err := client.CreateTable(&dynamodb.CreateTableInput{
@@ -125,88 +146,6 @@ func waitForTableToBeActiveWithRandomSleep(tableName string, client *dynamodb.Dy
 	return errors.WithStackTrace(TableActiveRetriesExceeded{TableName: tableName, Retries: maxRetries})
 }
 
-// Remove the given item from the DynamoDB lock table
-func removeItemFromLockTable(itemId string, tableName string, client *dynamodb.DynamoDB) error {
-	// TODO: should we check that the entry has our own metadata and not someone else's?
-
-	// Unless you specify conditions, the DeleteItem is an idempotent operation;
-	// running it multiple times on the same item or attribute does not result in an error response.
-	// https://docs.aws.amazon.com/sdk-for-go/api/service/dynamodb/#DynamoDB.DeleteItem
-	_, err := client.DeleteItem(&dynamodb.DeleteItemInput{
-		Key:                 createKeyFromItemId(itemId),
-		TableName:           aws.String(tableName),
-		ConditionExpression: aws.String(fmt.Sprintf("attribute_exists(%s)", ATTR_STATE_FILE_ID)),
-	})
-
-	// handle expected errors
-	if err != nil {
-		if awsErr, ok := err.(awserr.Error); ok {
-			switch awsErr.Code() {
-			case "ConditionalCheckFailedException":
-				err = ItemDoesNotExist{itemId, tableName, awsErr}
-			case "ResourceNotFoundException":
-				err = TableDoesNotExist{tableName, awsErr}
-			}
-		}
-	}
-
-	return errors.WithStackTrace(err)
-}
-
-// Write the given item to the DynamoDB lock table. If the given item already exists, return an error.
-func writeItemToLockTable(itemId string, tableName string, client *dynamodb.DynamoDB) error {
-	item, err := createItemAttributes(itemId, client)
-	if err != nil {
-		return err
-	}
-
-	// Conditional writes in DynamoDB should be strongly consistent: http://stackoverflow.com/a/23371813/483528
-	// https://r.32k.io/locking-with-dynamodb
-	_, err = client.PutItem(&dynamodb.PutItemInput{
-		TableName:           aws.String(tableName),
-		Item:                item,
-		ConditionExpression: aws.String(fmt.Sprintf("attribute_not_exists(%s)", ATTR_STATE_FILE_ID)),
-	})
-
-	return errors.WithStackTrace(err)
-}
-
-// Try to write the given item to the DynamoDB lock table. If the item already exists, that means someone already has
-// the lock, so display their metadata, sleep for the given amount of time, and try again, up to a maximum of
-// maxRetries retries.
-func writeItemToLockTableUntilSuccess(itemId string, tableName string, client *dynamodb.DynamoDB, maxRetries int, sleepBetweenRetries time.Duration, terragruntOptions *options.TerragruntOptions) error {
-	for retries := 1; ; retries++ {
-		terragruntOptions.Logger.Printf("Attempting to create lock item for state file %s in DynamoDB table %s", itemId, tableName)
-
-		err := writeItemToLockTable(itemId, tableName, client)
-		if err == nil {
-			terragruntOptions.Logger.Printf("Lock acquired!")
-			return nil
-		}
-
-		if !isItemAlreadyExistsErr(err) {
-			return err
-		}
-
-		displayLockMetadata(itemId, tableName, client, terragruntOptions)
-
-		if retries >= maxRetries {
-			return errors.WithStackTrace(AcquireLockRetriesExceeded{ItemId: itemId, Retries: maxRetries})
-		}
-
-		terragruntOptions.Logger.Printf("Will try to acquire lock again in %s.", sleepBetweenRetries)
-		time.Sleep(sleepBetweenRetries)
-	}
-}
-
-// Return true if the given error is the error returned by AWS when a conditional check fails. This is usually
-// indicates an item you tried to create already exists.
-func isItemAlreadyExistsErr(err error) bool {
-	unwrappedErr := errors.Unwrap(err)
-	awsErr, isAwsErr := unwrappedErr.(awserr.Error)
-	return isAwsErr && awsErr.Code() == "ConditionalCheckFailedException"
-}
-
 type TableActiveRetriesExceeded struct {
 	TableName string
 	Retries   int
@@ -216,15 +155,6 @@ func (err TableActiveRetriesExceeded) Error() string {
 	return fmt.Sprintf("Table %s is still not in active state after %d retries.", err.TableName, err.Retries)
 }
 
-type AcquireLockRetriesExceeded struct {
-	ItemId  string
-	Retries int
-}
-
-func (err AcquireLockRetriesExceeded) Error() string {
-	return fmt.Sprintf("Unable to acquire lock for item %s after %d retries.", err.ItemId, err.Retries)
-}
-
 type TableDoesNotExist struct {
 	TableName  string
 	Underlying error
@@ -232,14 +162,4 @@ type TableDoesNotExist struct {
 
 func (err TableDoesNotExist) Error() string {
 	return fmt.Sprintf("Table %s does not exist in DynamoDB! Original error from AWS: %v", err.TableName, err.Underlying)
-}
-
-type ItemDoesNotExist struct {
-	ItemId     string
-	TableName  string
-	Underlying error
-}
-
-func (err ItemDoesNotExist) Error() string {
-	return fmt.Sprintf("Item %s does not exist in %s DynamoDB table! Original error from AWS: %v", err.ItemId, err.TableName, err.Underlying)
 }
