--- test/fixture-lock/main.tf
@@ -1,4 +1,0 @@
-# Create an arbitrary local resource
-data "template_file" "test" {
-  template = "Hello, I am a template."
-}
