+++ test/fixture-modules/module-b/terraform.tfvars
@@ -1,12 +1,4 @@
 terragrunt = {
-  # Configure Terragrunt to use DynamoDB for locking
-  lock {
-    backend = "dynamodb"
-    config {
-      state_file_id = "${path_relative_to_include()}"
-    }
-  }
-
   # Configure Terragrunt to automatically store tfstate files in an S3 bucket
   remote_state {
     backend = "s3"
