+++ test/fixture-old-terragrunt-config/include-child-updated/.terragrunt
@@ -1,12 +1,3 @@
-# Configure Terragrunt to use DynamoDB for locking
-lock {
-  backend = "dynamodb"
-  config {
-    state_file_id = "${path_relative_to_include()}"
-    table_name = "terragrunt_locks_test_fixture_old_config_include_child_updated"
-  }
-}
-
 # Configure Terragrunt to automatically store tfstate files in an S3 bucket
 remote_state {
   backend = "s3"
