+++ test/fixture-old-terragrunt-config/include-parent-updated/terraform.tfvars
@@ -1,13 +1,4 @@
 terragrunt = {
-  # Configure Terragrunt to use DynamoDB for locking
-  lock {
-    backend = "dynamodb"
-    config {
-      state_file_id = "${path_relative_to_include()}"
-      table_name = "terragrunt_locks_test_fixture_old_config_include_parent_updated"
-    }
-  }
-
   # Configure Terragrunt to automatically store tfstate files in an S3 bucket
   remote_state {
     backend = "s3"
