--- test/fixture-old-terragrunt-config/lock/main.tf
@@ -1,3 +1,0 @@
-output "text" {
-  value = "Hello, World"
-}
